Added timeouts for all plugins that post.
All plugins now check to see if the user has called the comand in the
past 2 minutes. If so the bot returns a message about the cool-down
period. This is only in effect when the bot is posting to a channel and
not a direct message.

Added plugins/plugins.go to hold shared methods across all plugins.