Rule agreement command to auto set a chat role.
The idea is to allow a user to run an agreement command before able to
chat in the main channels of the Discord guild.

Moved AkSession to the `plugins.go` file.
Moved load function into its own package
Moved saved data files into a data directory and updated the
corresponding go files.